Move answer + guess derived selectors into Progress
