+++ bee-redux/src/features/puzzle/api/puzzleApiSlice.ts
@@ -13,161 +13,21 @@
 import { apiSlice } from "@/features/api";
 import { calculateScore } from "@/util";
 import { sortBy } from "lodash";
-
-export type Rank = {
-  id: string;
-  name: string;
-  multiplier: number;
-  rank: number;
-  score: number;
-};
-
-export type RanksType = [
+import {
   Rank,
-  Rank,
-  Rank,
-  Rank,
-  Rank,
-  Rank,
-  Rank,
-  Rank,
-  Rank,
-  Rank,
-];
-
-export const BlankRank: Rank = {
-  id: "noPuzzle",
-  name: "No Puzzle",
-  multiplier: 0,
-  rank: 0,
-  score: 0,
-};
-
-export const Ranks: RanksType = [
-  {
-    id: "beginner",
-    name: "Beginner",
-    multiplier: 0,
-    rank: 1,
-    score: 0,
-  },
-  {
-    id: "goodStart",
-    name: "Good Start",
-    multiplier: 0.02,
-    rank: 2,
-    score: 0,
-  },
-  {
-    id: "movingUp",
-    name: "Moving Up",
-    multiplier: 0.05,
-    rank: 3,
-    score: 0,
-  },
-  {
-    id: "good",
-    name: "Good",
-    multiplier: 0.08,
-    rank: 4,
-    score: 0,
-  },
-  {
-    id: "solid",
-    name: "Solid",
-    multiplier: 0.15,
-    rank: 5,
-    score: 0,
-  },
-  {
-    id: "nice",
-    name: "Nice",
-    multiplier: 0.25,
-    rank: 6,
-    score: 0,
-  },
-  {
-    id: "great",
-    name: "Great",
-    multiplier: 0.4,
-    rank: 7,
-    score: 0,
-  },
-  {
-    id: "amazing",
-    name: "Amazing",
-    multiplier: 0.5,
-    rank: 8,
-    score: 0,
-  },
-  {
-    id: "genius",
-    name: "Genius",
-    multiplier: 0.7,
-    rank: 9,
-    score: 0,
-  },
-  {
-    id: "queenBee",
-    name: "Queen Bee",
-    multiplier: 1,
-    rank: 10,
-    score: 0,
-  },
-];
-
-export type AnswerFormat = {
-  word: string;
-  frequency: number;
-  definitions: string[];
-};
-
-export type PuzzleFormatRaw = {
-  id: number;
-  date: string;
-  centerLetter: string;
-  outerLetters: string[];
-  validLetters: string[];
-  pangrams: string[];
-  perfectPangrams: string[];
-  answers: AnswerFormat[];
-  excludedWords: string[];
-  isLatest: boolean;
-};
-
-export type PuzzleFormat = PuzzleFormatRaw & {
-  shuffledOuterLetters: string[];
-  answerWords: string[];
-  totalPoints: number;
-  answerLengths: number[];
-  ranks: RanksType | [Rank];
-};
-
-export const BlankPuzzle: PuzzleFormat = {
-  id: 0,
-  date: "00-00-0000",
-  centerLetter: "_",
-  outerLetters: ["_", "_", "_", "_", "_", "_"],
-  shuffledOuterLetters: ["_", "_", "_", "_", "_", "_", "_"],
-  validLetters: ["_", "_", "_", "_", "_", "_", "_"],
-  pangrams: [],
-  perfectPangrams: [],
-  answers: [],
-  excludedWords: [],
-  isLatest: false,
-  answerWords: [],
-  totalPoints: 0,
-  answerLengths: [],
-  ranks: [BlankRank],
-};
+  Ranks,
+  RawPuzzle,
+  TPuzzle,
+  TRanks,
+} from "@/features/puzzle/types/puzzleTypes";
 
 export const puzzleApiSlice = apiSlice.injectEndpoints({
   endpoints: (builder) => ({
-    getPuzzle: builder.query<PuzzleFormat, string>({
+    getPuzzle: builder.query<TPuzzle, string>({
       query: (identifier: string) => ({
         url: `/puzzles/${identifier}`,
       }),
-      transformResponse: (response: PuzzleFormatRaw, _meta, arg) => {
+      transformResponse: (response: RawPuzzle, _meta, arg) => {
         const answerWords = response.answers.map((answer) => answer.word);
         const answerLengths = (answerWords: string[]) => {
           const returnArray: number[] = [];
@@ -180,7 +40,7 @@ export const puzzleApiSlice = apiSlice.injectEndpoints({
         };
         const totalPoints = calculateScore(answerWords);
 
-        const processedResponse: PuzzleFormat = {
+        const processedResponse: TPuzzle = {
           ...response,
           shuffledOuterLetters: [...response.outerLetters],
           answerWords,
@@ -191,7 +51,7 @@ export const puzzleApiSlice = apiSlice.injectEndpoints({
               ...rank,
               score: Math.round(rank.multiplier * totalPoints),
             }),
-          ) as RanksType,
+          ) as TRanks,
         };
         return processedResponse;
       },
