+++ bee-redux/src/features/puzzle/index.ts
@@ -3,3 +3,4 @@ export * from "./api/puzzleSlice";
 export * from "./components/Hive";
 export * from "./components/PuzzleNav";
 export * from "./hooks/useCurrentPuzzle";
+export * from "./types/puzzleTypes";
