+++ bee-redux/src/features/wordLists/components/knownWords/KnownWordsContainer.tsx
@@ -11,7 +11,6 @@
 */
 
 import { useAppDispatch, useAppSelector } from "@/app/hooks";
-import { selectCorrectGuesses } from "@/features/guesses";
 import {
   selectKnownWordsListSettings,
   setKnownWordsSortOrder,
@@ -22,6 +21,7 @@ import { KnownWordsStatus } from "./KnownWordsStatus";
 import { KnownWordsSettings } from "./KnownWordsSettings";
 import { getSortedGuessWordList } from "@/features/wordLists/util/wordListsUtil";
 import { WordListContainer } from "@/features/wordLists/components/WordListContainer";
+import { selectCorrectGuesses } from "@/features/progress/api/progressSelectors";
 
 export function KnownWordsContainer() {
   const dispatch = useAppDispatch();
